+++ types/network/entities.d.ts
@@ -3,6 +3,61 @@
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
+// The type of grantee:
+export type GranteeType =
+	// usr - Zimbra user
+	| 'usr'
+	// grp - Zimbra group(distribution list)
+	| 'grp'
+	// an external AD group
+	| 'egp'
+	// all - all authenticated users
+	| 'all'
+	// Zimbra Domain
+	| 'dom'
+	// non-Zimbra domain (used with sendToDistList right)
+	| 'edom'
+	// gst - non-Zimbra email address and password (not yet supported)
+	| 'gst'
+	// key - external user with an accesskey
+	| 'key'
+	// pub - public authenticated and unauthenticated access
+	| 'pub'
+	// Pseudo grantee type.  Granting code will map to usr/grp/egp or gst
+	| 'email';
+
+export type Right =
+	// Following rights are partial, they are the result of
+	// description: automatically add meeting invites from grantee to the target's calendar
+	// right type: preset
+	// target type(s): account
+	// grant target type: (default)
+	// right class: USER
+	| 'invite'
+	// description: login as another user.  Currently this is only honored for imap/pop3 login.
+	// right type: preset
+	// target type(s): account
+	// grant target type: (default)
+	// right class: USER
+	| 'loginAs'
+	// description: reply to messages in a shared folder as the owner of the folder
+	// right type: preset
+	// target type(s): account
+	// grant target type: account
+	// right class: USER
+	| 'sendAs'
+	// description: send messages on behalf of the grantor
+	// right type: preset
+	// target type(s): account
+	// grant target type: account
+	// right class: USER
+	| 'sendOnBehalfOf'
+	// description: view free/busy
+	// right type: preset
+	// target type(s): account
+	// grant target type: (default)
+	// right class: USER
+	| 'viewFreeBusy';
 
 /** Specify Access Control Entries */
 export interface AccountACEInfo {
@@ -32,9 +87,9 @@ export interface AccountACEInfo {
 	 * if key is not given, server will generate an access key
 	 * If chkgt is "1 (true)", INVALID_REQUEST will be thrown if wrong grantee type is specified.
 	 */
-	gt: 'usr' | 'grp' | 'egp' | 'all' | 'dom' | 'edom' | 'gst' | 'key' | 'pub' | 'email';
+	gt: GranteeType;
 	/** Right */
-	right: 'viewFreeBusy' | 'invite';
+	right: Right;
 	/**
 	 * Name or email address of the grantee.
 	 * Not present if {grantee-type} is "all" or "pub"
