+++ types/network/index.d.ts
@@ -232,60 +232,8 @@ export type GetRightsRequest = {
 	_jsns: NameSpace;
 };
 
-export type Right =
-	// Following rights are partial, they are the result of
-	// description: automatically add meeting invites from grantee to the target's calendar
-	// right type: preset
-	// target type(s): account
-	// grant target type: (default)
-	// right class: USER
-	| 'invite'
-	// description: login as another user.  Currently this is only honored for imap/pop3 login.
-	// right type: preset
-	// target type(s): account
-	// grant target type: (default)
-	// right class: USER
-	| 'loginAs'
-	// description: reply to messages in a shared folder as the owner of the folder
-	// right type: preset
-	// target type(s): account
-	// grant target type: account
-	// right class: USER
-	| 'sendAs'
-	// description: send messages on behalf of the grantor
-	// right type: preset
-	// target type(s): account
-	// grant target type: account
-	// right class: USER
-	| 'sendOnBehalfOf'
-	// description: view free/busy
-	// right type: preset
-	// target type(s): account
-	// grant target type: (default)
-	// right class: USER
-	| 'viewFreeBusy';
-
 export type NameSpace = 'urn:zimbraMail' | 'urn:zimbraAccount' | 'urn:zimbra';
 
-// The type of grantee:
-export type GranteeType =
-	// usr - Zimbra user
-	| 'usr'
-	// grp - Zimbra group(distribution list)
-	| 'grp'
-	// an external AD group
-	| 'egp'
-	// all - all authenticated users
-	| 'all'
-	// Zimbra Domain
-	| 'dom'
-	// non-Zimbra domain (used with sendToDistList right)
-	| 'edom'
-	// gst - non-Zimbra email address and password (not yet supported)
-	| 'gst'
-	// key - external user with an accesskey
-	| 'key'
-	// pub - public authenticated and unauthenticated access
-	| 'pub'
-	// Pseudo grantee type.  Granting code will map to usr/grp/egp or gst
-	| 'email';
+export type GetRightsResponse = {
+	ace?: Array<AccountACEInfo>;
+};
